etl/postgres_to_es/requirements.txt
elasticsearch==8.11.1
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
flake8-isort==6.1.1
redis==5.0.1
backoff==2.2.1
ruff==0.2.1